+++ CHANGELOG.md
@@ -17,6 +17,8 @@
   - Individual variable pages have been redesigned and include additional information about known issues and variable history
   - Folder navigation has been simplified on the variable index
   - Embedded PDFs for variables are now displayed directly on the variable page
+- **Dashboard Added**
+  - Users can now view their personal dashboard that contains updates and links to their datasets and other NSRR related activity
 - **Blog Added**
   - Community managers can now create new blogs posts that are viewable on the NSRR home page
 - **Gem Changes**